Refactor logging to use c_str() for string conversion
Updated logging statements to use c_str() for string conversion in various locations for consistency and clarity.

Got a little heavy handed and removed too many of these in the last edit. Should have just been this->parent_->address_str().c_str() as in 2025.12.0 update, ESPHome changed the address_str() function. It used to return a "String object" (which needed .c_str() to work), but now it returns a "Raw String" directly.